refactor!: new implementation (#77)
