+++ index.js
@@ -13,21 +13,7 @@
  */
 const isUtf8MimeType = require('./lib/isUtf8MimeType').isUtf8MimeType
 const mime = require('mime')
-const SendStream = require('./lib/SendStream')
-
-/**
- * Return a `SendStream` for `req` and `path`.
- *
- * @param {object} req
- * @param {string} path
- * @param {object} [options]
- * @return {SendStream}
- * @public
- */
-
-function send (req, path, options) {
-  return new SendStream(req, path, options)
-}
+const send = require('./lib/send').send
 
 /**
  * Module exports.
@@ -37,7 +23,6 @@ function send (req, path, options) {
 module.exports = send
 module.exports.default = send
 module.exports.send = send
-module.exports.SendStream = SendStream
 
 module.exports.isUtf8MimeType = isUtf8MimeType
 module.exports.mime = mime
