--- lib/setHeaders.js
@@ -1,20 +1,0 @@
-'use strict'
-
-/**
- * Set an object of headers on a response.
- *
- * @param {object} res
- * @param {object} headers
- * @private
- */
-
-function setHeaders (res, headers) {
-  const keys = Object.keys(headers)
-
-  for (let i = 0; i < keys.length; i++) {
-    const key = keys[i]
-    res.setHeader(key, headers[key])
-  }
-}
-
-module.exports.setHeaders = setHeaders
