+++ package.json
@@ -24,7 +24,6 @@
   "dependencies": {
     "escape-html": "~1.0.3",
     "fast-decode-uri-component": "^1.0.1",
-    "http-errors": "2.0.0",
     "mime": "^3",
     "@lukeed/ms": "^2.0.2"
   },
