+++ test/send.1.test.js
@@ -5,14 +5,14 @@ const fs = require('node:fs')
 const http = require('node:http')
 const path = require('node:path')
 const request = require('supertest')
-const SendStream = require('..').SendStream
+const send = require('..').send
 const { shouldNotHaveHeader, createServer } = require('./utils')
 
 // test server
 
 const fixtures = path.join(__dirname, 'fixtures')
 
-test('SendStream(file, options)', function (t) {
+test('send(file, options)', function (t) {
   t.plan(10)
 
   t.test('acceptRanges', function (t) {
@@ -184,14 +184,14 @@ test('SendStream(file, options)', function (t) {
     t.test('when "allow"', function (t) {
       t.plan(3)
 
-      t.test('should SendStream dotfile', function (t) {
+      t.test('should send dotfile', function (t) {
         t.plan(1)
         request(createServer({ dotfiles: 'allow', root: fixtures }))
           .get('/.hidden.txt')
           .expect(200, 'secret', err => t.error(err))
       })
 
-      t.test('should SendStream within dotfile directory', function (t) {
+      t.test('should send within dotfile directory', function (t) {
         t.plan(1)
         request(createServer({ dotfiles: 'allow', root: fixtures }))
           .get('/.mine/name.txt')
@@ -265,7 +265,7 @@ test('SendStream(file, options)', function (t) {
           .expect(403, err => t.error(err))
       })
 
-      t.test('should SendStream files in root dotfile directory', function (t) {
+      t.test('should send files in root dotfile directory', function (t) {
         t.plan(1)
         request(createServer({ dotfiles: 'deny', root: path.join(fixtures, '.mine') }))
           .get('/name.txt')
@@ -274,8 +274,10 @@ test('SendStream(file, options)', function (t) {
 
       t.test('should 403 for dotfile without root', function (t) {
         t.plan(1)
-        const server = http.createServer(function onRequest (req, res) {
-          new SendStream(req, fixtures + '/.mine' + req.url, { dotfiles: 'deny' }).pipe(res)
+        const server = http.createServer(async function onRequest (req, res) {
+          const { statusCode, headers, stream } = await send(req, fixtures + '/.mine' + req.url, { dotfiles: 'deny' })
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(server)
@@ -335,7 +337,7 @@ test('SendStream(file, options)', function (t) {
           .expect(404, err => t.error(err))
       })
 
-      t.test('should SendStream files in root dotfile directory', function (t) {
+      t.test('should send files in root dotfile directory', function (t) {
         t.plan(1)
 
         request(createServer({ dotfiles: 'ignore', root: path.join(fixtures, '.mine') }))
@@ -346,8 +348,10 @@ test('SendStream(file, options)', function (t) {
       t.test('should 404 for dotfile without root', function (t) {
         t.plan(1)
 
-        const server = http.createServer(function onRequest (req, res) {
-          new SendStream(req, fixtures + '/.mine' + req.url, { dotfiles: 'ignore' }).pipe(res)
+        const server = http.createServer(async function onRequest (req, res) {
+          const { statusCode, headers, stream } = await send(req, fixtures + '/.mine' + req.url, { dotfiles: 'ignore' })
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(server)
@@ -487,10 +491,11 @@ test('SendStream(file, options)', function (t) {
     t.test('should work without root', function (t) {
       t.plan(1)
 
-      const server = http.createServer(function (req, res) {
+      const server = http.createServer(async function (req, res) {
         const p = path.join(fixtures, 'pets').replace(/\\/g, '/') + '/'
-        new SendStream(req, p, { index: ['index.html'] })
-          .pipe(res)
+        const { statusCode, headers, stream } = await send(req, p, { index: ['index.html'] })
+        res.writeHead(statusCode, headers)
+        stream.pipe(res)
       })
 
       request(server)
@@ -515,9 +520,10 @@ test('SendStream(file, options)', function (t) {
       t.test('should work with trailing slash', function (t) {
         t.plan(1)
 
-        const app = http.createServer(function (req, res) {
-          new SendStream(req, req.url, { root: fixtures + '/' })
-            .pipe(res)
+        const app = http.createServer(async function (req, res) {
+          const { statusCode, headers, stream } = await send(req, req.url, { root: fixtures + '/' })
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(app)
@@ -528,9 +534,10 @@ test('SendStream(file, options)', function (t) {
       t.test('should work with empty path', function (t) {
         t.plan(1)
 
-        const app = http.createServer(function (req, res) {
-          new SendStream(req, '', { root: fixtures })
-            .pipe(res)
+        const app = http.createServer(async function (req, res) {
+          const { statusCode, headers, stream } = await send(req, '', { root: fixtures })
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(app)
@@ -547,9 +554,10 @@ test('SendStream(file, options)', function (t) {
       t.test('should try as file with empty path', function (t) {
         t.plan(1)
 
-        const app = http.createServer(function (req, res) {
-          new SendStream(req, '', { root: path.join(fixtures, 'name.txt') })
-            .pipe(res)
+        const app = http.createServer(async function (req, res) {
+          const { statusCode, headers, stream } = await send(req, '', { root: path.join(fixtures, 'name.txt') })
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(app)
@@ -561,20 +569,21 @@ test('SendStream(file, options)', function (t) {
         t.plan(1)
 
         request(createServer({ root: fixtures }))
-          .get('/pets/../../SendStream.js')
+          .get('/pets/../../send.js')
           .expect(403, err => t.error(err))
       })
 
       t.test('should allow .. in root', function (t) {
         t.plan(1)
 
-        const app = http.createServer(function (req, res) {
-          new SendStream(req, req.url, { root: fixtures + '/../fixtures' })
-            .pipe(res)
+        const app = http.createServer(async function (req, res) {
+          const { statusCode, headers, stream } = await send(req, req.url, { root: fixtures + '/../fixtures' })
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(app)
-          .get('/pets/../../SendStream.js')
+          .get('/pets/../../send.js')
           .expect(403, err => t.error(err))
       })
 
@@ -601,22 +610,24 @@ test('SendStream(file, options)', function (t) {
       t.test('should consider .. malicious', function (t) {
         t.plan(1)
 
-        const app = http.createServer(function (req, res) {
-          new SendStream(req, fixtures + req.url)
-            .pipe(res)
+        const app = http.createServer(async function (req, res) {
+          const { statusCode, headers, stream } = await send(req, fixtures + req.url)
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(app)
-          .get('/../SendStream.js')
+          .get('/../send.js')
           .expect(403, err => t.error(err))
       })
 
       t.test('should still serve files with dots in name', function (t) {
         t.plan(1)
 
-        const app = http.createServer(function (req, res) {
-          new SendStream(req, fixtures + req.url)
-            .pipe(res)
+        const app = http.createServer(async function (req, res) {
+          const { statusCode, headers, stream } = await send(req, fixtures + req.url)
+          res.writeHead(statusCode, headers)
+          stream.pipe(res)
         })
 
         request(app)
