+++ test/utils.js
@@ -10,10 +10,12 @@ module.exports.shouldNotHaveHeader = function shouldNotHaveHeader (header, t) {
 }
 
 module.exports.createServer = function createServer (opts, fn) {
-  return http.createServer(function onRequest (req, res) {
+  return http.createServer(async function onRequest (req, res) {
     try {
       fn && fn(req, res)
-      send(req, req.url, opts).pipe(res)
+      const { statusCode, headers, stream } = await send(req, req.url, opts)
+      res.writeHead(statusCode, headers)
+      stream.pipe(res)
     } catch (err) {
       res.statusCode = 500
       res.end(String(err))
